server/pyproject.toml
[project]
name = "bazaaryar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "fastapi[standard]>=0.128.0",
    "jdatetime>=5.2.0",
    "langchain>=1.2.8",
    "langchain-google-genai>=4.2.0",
    "langchain-openai==1.1.7",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.6.2",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.2.0",
    "pytest>=8.4.0",
]
